app-streamlit/requirements.txt
black
cartopy
catboost
click
dash-ag-grid
gensim
graphviz
ipython
joblib
lightgbm
linkedin-jobs-scraper
matplotlib
mypy
nltk
np_utils
numpy
ollama
opencv-python
openai
openpyxl
pandas
panel
pdf2image
pdfminer
pikepdf
plotly
pre-commit
psycopg2-binary
pycountry
pypdf
pytest
python-dotenv
python-pptx
pyyaml
requests
scikit-learn
seaborn
sentence-transformers
spacy
sqlalchemy
st-pages
streamlit
streamlit-aggrid
streamlit-feedback
streamlit_option_menu
streamlit-tags
torch
torchaudio
torchvision
transformers
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
vecstack
watchdog
wordcloud
xgboost